new plot: displot for node/edge attributes
